==> 1 - preparation_donnees.ipynb <==
import pandas as pd

df_PS = pd.read_csv("src/PS_LibreAcces_Personne_activite_202502190859.txt", sep="|", dtype=str)
df_communes = pd.read_csv("src/communes-france-2025.csv", sep=",", dtype=str)
df_demo_PS = pd.read_csv("src/demographie-effectifs-et-les-densites(1).csv", sep=";",dtype=str)

df_demo_PS=df_demo_PS[df_demo_PS['﻿annee']=='2023']
df_demo_PS=df_demo_PS[df_demo_PS['libelle_sexe']=='tout sexe']
df_demo_PS=df_demo_PS[~df_demo_PS['libelle_classe_age'].isin(('Tout âge','âge inconnu'))]

ON VEUT PAS LES DOUBLONS ET ON S'INTERESSE QU'AUX CIVILS

df_PS=df_PS.drop_duplicates(subset='Identifiant PP')
df_PS = df_PS[df_PS['Code catégorie professionnelle']=='C']

SUPPRESSION DES COLONNES NON-PERTINENTES

cols_communes_a_supprimer = ['Unnamed: 0',
                            'nom_sans_pronom', 
                            'nom_a',
                            'nom_de', 
                            'nom_sans_accent', 
                            'nom_standard_majuscule',
                            'typecom',
                            'typecom_texte',
                            'canton_code', 
                            'canton_nom', 
                            'epci_code', 
                            'epci_nom', 
                            'academie_code',
                            'academie_nom',
                            'codes_postaux', 
                            'zone_emploi',
                            'code_insee_centre_zone_emploi', 
                            'code_unite_urbaine',
                            'nom_unite_urbaine', 
                            'taille_unite_urbaine',
                            'type_commune_unite_urbaine', 
                            'statut_commune_unite_urbaine',
                            'gentile', 
                            'url_wikipedia', 
                            'url_villedereve',
                            'niveau_equipements_services', 
                            'niveau_equipements_services_texte']


cols_PS_a_supprimer =      ['Type d\'identifiant PP', 
                            'Identification nationale PP',  
                            'Code civilité d\'exercice',	
                            'Libellé civilité d\'exercice',
                            'Code catégorie professionnelle',
                            'Libellé catégorie professionnelle', 
                            'Code type savoir-faire',
                            'Libellé type savoir-faire', 
                            'Numéro SIRET site', 'Numéro SIREN site', 
                            'Numéro FINESS site',
                            'Numéro FINESS établissement juridique',
                            'Identifiant technique de la structure', 
                            'Enseigne commerciale site',
                            'Complément destinataire (coord. structure)',
                            'Complément point géographique (coord. structure)',
                            'Indice répétition voie (coord. structure)',
                            'Code type de voie (coord. structure)',
                            'Libellé type de voie (coord. structure)',
                            'Mention distribution (coord. structure)',
                            'Bureau cedex (coord. structure)', 
                            'Code pays (coord. structure)', 
                            'Libellé pays (coord. structure)',
                            'Téléphone 2 (coord. structure)',
                            'Télécopie (coord. structure)', 
                            'Code Département (structure)',
                            'Libellé Département (structure)',
                            'Ancien identifiant de la structure', 
                            'Autorité d\'enregistrement',
                            'Code section tableau pharmaciens',
                            'Libellé section tableau pharmaciens', 
                            'Code rôle', 
                            'Libellé rôle',
                            'Code genre activité', 
                            'Libellé genre activité', 
                            'Unnamed: 56']

cols_demographie_a_supprimer=['densite', 'vision generale all',
       'vision_generale_prescriptions', 'vision profession territoire']

df_communes=df_communes.drop(columns=cols_communes_a_supprimer)
df_PS = df_PS.drop(columns=cols_PS_a_supprimer)
df_demo_PS=df_demo_PS.drop(columns=cols_demographie_a_supprimer)

df_communes.to_csv('csv/communes.csv', sep=";",index=False)
df_PS.to_csv('csv/PS_civils.csv', sep=";",index=False)
df_demo_PS.to_csv('csv/demographie_ps.csv', sep=';',index=False)
