==> 5 - coordonnees_ps.ipynb <==
import plotly.express as px
import pandas as pd
import numpy as np
import seaborn as sns
import datetime
import requests
import random
from dotenv import load_dotenv   #for python-dotenv method
load_dotenv()                    #for python-dotenv method
import os

df_ps=pd.read_csv("csv/PS_civils.csv", sep=";", dtype=str)
df_sample = df_ps

df_sample=df_ps[~df_ps['Libellé Voie (coord. structure)'].isna()]

df_sample=df_sample[df_sample["Libellé profession"]=="Médecin"]

df_sample=df_sample[df_sample['Libellé savoir-faire'].isin(('Qualifié en Médecine Générale','Spécialiste en Médecine Générale','Médecine Générale'))]

df_sample=df_sample[df_sample['Libellé mode exercice']!="Salarié"]

df_sample=df_sample.sample(1000)

nominatim_user_agent = os.environ.get('nominatim_user_agent')

url = "https://nominatim.openstreetmap.org/search?"
headers = {'user_agent':nominatim_user_agent}

df_idf_avec_coordonnees=pd.DataFrame(columns=("insee", "code_postal", "nom", "lat", "lon"))
lats = []
longs = []
for i in range(0,len(df_sample)):
    
    street = f"{df_sample.iloc[i,13]} {df_sample.iloc[i,12]}"
    postal_code = df_sample.iloc[i,14]
    p = {'street':{street},'postalcode' : {postal_code},'country':'France', 'format':'json', 'limit':1, 'email':'charles@gmail.com'}
    r = requests.get(url, headers=headers, params=p)

    lat = None
    lon = None
    try:

        if len(r.json())>0:
                
            lats.append(r.json()[0]['lat'])
            longs.append(r.json()[0]['lon'])
        else:
            lats.append(None)
            longs.append(None)
    except:
        print("Error from server: " + str(r.content))



    print(f'{i}/{len(df_sample)} ({round(i/len(df_sample)*100,2)}%)')
    print('*************************')
df_sample.insert(loc=len(df_sample.columns), column='lat', value=lats)
df_sample.insert(loc=len(df_sample.columns), column='lon', value=longs)

df_sample.to_csv('csv/PS_civils_1000_avec_coordonnees.csv', sep=";",index=False)
